refactor: refact chart option
